refactor : 이미지 생성 컨트롤러 반환 값 수정
[목적]:  이미지 생성 컨트롤러 반환 값 수정

[목표]:
- 이미지 생성 컨트롤러 반환 값 수정

[달성도]:
- 이미지 생성 컨트롤러 반환 값 수정

[기타]:
